Major breakout of window code to Window.js
